Add base LayerOpMode class
